Add undocumented "meteor watch" command.
This is so that we can use some of the watch functionality in other
repositories. I would like to put watch into a package and load it using
`unipackage.load`, but there's a circular dependency (you need a library
to load a unipackage, and you need a WatchSet to create a library) and
this seems like the simplest way forward for now (especially because any
refactoring will probably conflict with tool-refactoring).